+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-12-31
+
+## Added
+- Support for ModularAvatarMergeAnimator (#9)
+
+### Fixed
+- Support for material variant (#11)
+- Ambient light in the Lighting window does not work in unity 2022 (#13)
+
 ## [1.0.3] - 2023-04-01
 
 ### Fixed